[v0.87.1][maintainability] Split demo.rs by responsibility
