+++ adl/src/demo/fixtures/demo_g_readme.md
@@ -1,0 +1,32 @@
+# D1 Canonical Bounded Cognitive Path
+
+Generated by:
+
+```bash
+cargo run --manifest-path adl/Cargo.toml --bin adl -- demo demo-g-v086-control-path --run --trace --out ./out
+```
+
+This demo emits one stable, reviewer-facing control-path artifact set covering:
+
+- signals
+- candidate selection
+- arbitration
+- bounded execution
+- evaluation
+- reframing
+- memory participation
+- Freedom Gate
+- final result
+
+Primary artifacts:
+- `signals.json`
+- `candidate_selection.json`
+- `arbitration.json`
+- `execution_iterations.json`
+- `evaluation.json`
+- `reframing.json`
+- `memory.json`
+- `freedom_gate.json`
+- `final_result.json`
+- `summary.txt`
+- `trace.jsonl`
