+++ adl/src/demo/obsmem.rs
@@ -1,0 +1,183 @@
+use std::path::{Path, PathBuf};
+
+use anyhow::{Context, Result};
+
+use crate::godel;
+
+use super::write_file;
+
+pub(super) const DEMO_D_FAILURE_SIGNAL_JSON: &str = r#"{
+  "schema_version": "godel_failure_signal.v1",
+  "run_id": "demo-d-run-001",
+  "workflow_id": "godel-obsmem-demo",
+  "failure_code": "tool_failure",
+  "failure_summary": "bounded deterministic failure signal for demo",
+  "evidence_refs": [
+    "runs/demo-d-run-001/logs/activation_log.json",
+    "runs/demo-d-run-001/run_status.json"
+  ]
+}
+"#;
+
+pub(super) const DEMO_D_README_MD: &str = r#"# Demo D Output — Gödel + ObsMem Loop
+
+Generated by:
+
+```bash
+cargo run --manifest-path adl/Cargo.toml --bin adl -- demo demo-d-godel-obsmem-loop --run --trace --out ./out
+```
+
+This demo executes the bounded Gödel stage loop and persists:
+
+- `runs/demo-d-run-001/godel/canonical_evidence_view.v1.json`
+- `runs/demo-d-run-001/godel/experiment_record.runtime.v1.json`
+- `runs/demo-d-run-001/godel/obsmem_index_entry.runtime.v1.json`
+- `godel_obsmem_demo_summary.json`
+- `trace.jsonl`
+"#;
+
+pub(super) const DEMO_F_README_MD: &str = r#"# Demo F Output — ObsMem Retrieval
+
+Generated by:
+
+```bash
+cargo run --manifest-path adl/Cargo.toml --bin adl -- demo demo-f-obsmem-retrieval --run --trace --out ./out
+```
+
+This demo seeds deterministic runtime experiment/index artifacts and performs deterministic retrieval by:
+
+- `failure_code`
+- optional `hypothesis_id`
+- optional `experiment_outcome`
+
+Primary artifacts:
+- `runs/demo-f-run-a/godel/experiment_record.runtime.v1.json`
+- `runs/demo-f-run-a/godel/obsmem_index_entry.runtime.v1.json`
+- `runs/demo-f-run-b/godel/experiment_record.runtime.v1.json`
+- `runs/demo-f-run-b/godel/obsmem_index_entry.runtime.v1.json`
+- `obsmem_retrieval_result.json`
+- `trace.jsonl`
+"#;
+
+pub(super) fn run_godel_stage_loop_demo(out_dir: &Path) -> Result<Vec<PathBuf>> {
+    let mut artifacts = Vec::new();
+    let runs_root = out_dir.join("runs");
+    let exec = godel::GodelStageLoopExecutor::new(godel::StageLoopConfig::default());
+    let input = godel::StageLoopInput {
+        run_id: "demo-d-run-001".to_string(),
+        workflow_id: "godel-obsmem-demo".to_string(),
+        failure_code: "tool_failure".to_string(),
+        failure_summary: "bounded deterministic failure signal for demo".to_string(),
+        evidence_refs: vec![
+            "runs/demo-d-run-001/logs/activation_log.json".to_string(),
+            "runs/demo-d-run-001/run_status.json".to_string(),
+        ],
+    };
+    let persisted = exec.execute_and_persist(&input, &runs_root)?;
+    let summary = serde_json::json!({
+        "schema_version": "godel_obsmem_demo_summary.v1",
+        "run_id": persisted.run.record.run_id,
+        "stage_order": persisted
+            .run
+            .stage_order
+            .iter()
+            .map(|s| s.as_str())
+            .collect::<Vec<_>>(),
+        "hypothesis_count": persisted.run.hypotheses.len(),
+        "selected_hypothesis_id": persisted.run.hypothesis.id,
+        "selected_mutation_id": persisted.run.mutation.id,
+        "canonical_mutation_rel_path": persisted.canonical_mutation_rel_path,
+        "evaluation_decision": format!("{:?}", persisted.run.evaluation.decision).to_lowercase(),
+        "canonical_evaluation_plan_rel_path": persisted.canonical_evaluation_plan_rel_path,
+        "canonical_evidence_rel_path": persisted.canonical_evidence_rel_path,
+        "experiment_record_rel_path": persisted.experiment_record_rel_path,
+        "obsmem_index_rel_path": persisted.obsmem_index_rel_path
+    });
+    artifacts.push(write_file(
+        out_dir,
+        "godel_obsmem_demo_summary.json",
+        &serde_json::to_string_pretty(&summary)?,
+    )?);
+    artifacts.push(out_dir.join("runs/demo-d-run-001/godel/evaluation_plan.v1.json"));
+    artifacts.push(out_dir.join("runs/demo-d-run-001/godel/mutation.v1.json"));
+    artifacts.push(out_dir.join("runs/demo-d-run-001/godel/canonical_evidence_view.v1.json"));
+    artifacts.push(out_dir.join("runs/demo-d-run-001/godel/evaluation_plan.v1.json"));
+    artifacts.push(out_dir.join("runs/demo-d-run-001/godel/experiment_record.runtime.v1.json"));
+    artifacts.push(out_dir.join("runs/demo-d-run-001/godel/obsmem_index_entry.runtime.v1.json"));
+    Ok(artifacts)
+}
+
+pub(super) fn seed_obsmem_retrieval_demo(out_dir: &Path) -> Result<Vec<PathBuf>> {
+    let mut artifacts = Vec::new();
+    let runs_root = out_dir.join("runs");
+    let exec = godel::GodelStageLoopExecutor::new(godel::StageLoopConfig::default());
+
+    let input_a = godel::StageLoopInput {
+        run_id: "demo-f-run-a".to_string(),
+        workflow_id: "godel-retrieval-demo".to_string(),
+        failure_code: "tool_failure".to_string(),
+        failure_summary: "deterministic failure A".to_string(),
+        evidence_refs: vec!["runs/demo-f-run-a/run_status.json".to_string()],
+    };
+    let mut input_b = input_a.clone();
+    input_b.run_id = "demo-f-run-b".to_string();
+    input_b.failure_code = "policy_denied".to_string();
+    input_b.failure_summary = "deterministic failure B".to_string();
+    input_b.evidence_refs = vec!["runs/demo-f-run-b/run_status.json".to_string()];
+
+    for input in [input_a, input_b] {
+        let persisted = exec.execute_and_persist(&input, &runs_root)?;
+        artifacts.push(out_dir.join(&persisted.experiment_record_rel_path));
+        artifacts.push(out_dir.join(&persisted.obsmem_index_rel_path));
+    }
+
+    Ok(artifacts)
+}
+
+pub(super) fn query_obsmem_retrieval_demo(out_dir: &Path) -> Result<Vec<PathBuf>> {
+    let mut entries: Vec<godel::obsmem_index::StageIndexEntry> = Vec::new();
+    let runs_root = out_dir.join("runs");
+    for run_id in ["demo-f-run-a", "demo-f-run-b"] {
+        let path = runs_root
+            .join(run_id)
+            .join("godel")
+            .join("obsmem_index_entry.runtime.v1.json");
+        let raw = std::fs::read_to_string(&path)
+            .with_context(|| format!("failed to read '{}'", path.display()))?;
+        let persisted: godel::obsmem_index::PersistedStageIndexEntry =
+            serde_json::from_str(&raw)
+                .with_context(|| format!("failed to parse '{}'", path.display()))?;
+        entries.push(persisted.entry);
+    }
+
+    let query = godel::obsmem_index::ObsMemIndexQuery {
+        failure_code: "tool_failure".to_string(),
+        hypothesis_id: None,
+        experiment_outcome: None,
+    };
+
+    let mut matches: Vec<_> = entries
+        .into_iter()
+        .filter(|e| godel::obsmem_index::matches_query(e, &query))
+        .collect();
+    matches.sort_by(|a, b| {
+        a.index_key
+            .cmp(&b.index_key)
+            .then(a.run_id.cmp(&b.run_id))
+            .then(a.mutation_id.cmp(&b.mutation_id))
+    });
+
+    let result = serde_json::json!({
+        "schema_version": "obsmem_retrieval_result.demo.v1",
+        "query": query,
+        "result_count": matches.len(),
+        "results": matches,
+    });
+
+    let path = write_file(
+        out_dir,
+        "obsmem_retrieval_result.json",
+        &serde_json::to_string_pretty(&result)?,
+    )?;
+    Ok(vec![path])
+}
