+++ adl/src/demo/pipeline.rs
@@ -1,0 +1,85 @@
+use crate::prompt;
+
+pub(super) const DEMO_E_INPUT_CARD_MD: &str = r#"# Input Card (Demo Fixture)
+
+Task: produce a concise release-note paragraph for a deterministic CLI improvement.
+
+Prompt Spec:
+- actor: writer
+- model: bounded-demo
+- outputs: markdown paragraph
+"#;
+
+pub(super) const DEMO_E_WRITER_MD: &str = r#"# Stage 1 — Writer
+
+The CLI now includes deterministic demo surfaces with explicit artifact paths and stable replay-oriented output.
+"#;
+
+pub(super) const DEMO_E_EDITOR_MD: &str = r#"# Stage 2 — Editor
+
+Edited for clarity: each demo documents command, artifacts, and deterministic constraints.
+"#;
+
+pub(super) const DEMO_E_COPYEDITOR_MD: &str = r#"# Stage 3 — Copyeditor
+
+Copyedited to remove ambiguity and keep all artifact paths repo-relative.
+"#;
+
+pub(super) const DEMO_E_README_MD: &str = r#"# Demo E Output — Multi-Agent Card Pipeline
+
+Generated by:
+
+```bash
+cargo run --manifest-path adl/Cargo.toml --bin adl -- demo demo-e-multi-agent-card-pipeline --run --trace --out ./out
+```
+
+Stages:
+- writer -> editor -> copyeditor -> publisher
+
+Primary artifacts:
+- `pipeline/input_card.md`
+- `pipeline/01_writer.md`
+- `pipeline/02_editor.md`
+- `pipeline/03_copyeditor.md`
+- `pipeline/pipeline_manifest.json`
+- `trace.jsonl`
+"#;
+
+#[derive(Debug, serde::Serialize)]
+pub(super) struct CardPipelineManifest {
+    schema_version: &'static str,
+    stage_order: Vec<&'static str>,
+    stage_artifacts: Vec<CardPipelineStageArtifact>,
+}
+
+#[derive(Debug, serde::Serialize)]
+struct CardPipelineStageArtifact {
+    stage: &'static str,
+    path: &'static str,
+    content_hash: String,
+}
+
+pub(super) fn build_card_pipeline_manifest() -> CardPipelineManifest {
+    let stage_artifacts = vec![
+        ("writer", "pipeline/01_writer.md", DEMO_E_WRITER_MD),
+        ("editor", "pipeline/02_editor.md", DEMO_E_EDITOR_MD),
+        (
+            "copyeditor",
+            "pipeline/03_copyeditor.md",
+            DEMO_E_COPYEDITOR_MD,
+        ),
+    ]
+    .into_iter()
+    .map(|(stage, path, content)| CardPipelineStageArtifact {
+        stage,
+        path,
+        content_hash: prompt::hash_prompt(content),
+    })
+    .collect();
+
+    CardPipelineManifest {
+        schema_version: "multi_agent_card_pipeline.demo.v1",
+        stage_order: vec!["writer", "editor", "copyeditor", "publisher"],
+        stage_artifacts,
+    }
+}
